+++ generatedComponentConfig.js
@@ -20,7 +20,6 @@ import TerraThemeProviderIndex from './packages/terra-theme-provider/examples/In
 import AggregatorStandalone from './packages/terra-aggregator/examples/test-examples/AggregatorStandalone.example';
 import AggregatorWithDisclosure from './packages/terra-aggregator/examples/test-examples/AggregatorWithDisclosure.example';
 import ApplicationHeaderDefault from './packages/terra-application-header-layout/examples/test-examples/ApplicationHeaderDefault.example';
-import ApplicationListDefault from './packages/terra-application-links/examples/test-examples/ApplicationListDefault.example';
 import ApplicationTabsDefault from './packages/terra-application-links/examples/test-examples/ApplicationTabsDefault.example';
 import ApplicationMenuDefault from './packages/terra-application-menu-layout/examples/test-examples/ApplicationMenuDefault.example';
 import ApplicationHeaderNameDefault from './packages/terra-application-name/examples/test-examples/ApplicationHeaderNameDefault.example';
@@ -108,11 +107,6 @@ const componentConfig = {
       },
     ],
     tests: [
-      {
-        name: 'Application List Default',
-        path: '/application-list-default',
-        component: ApplicationListDefault,
-      },
       {
         name: 'Application Tabs Default',
         path: '/application-tabs-default',