+++ packages/terra-application-links/CHANGELOG.md
@@ -9,6 +9,7 @@ Unreleased
 ### Changed
 * Code formatting and documentation spacing.
 * Updated screenshots tests to match new themeable variables in terra-core.
+* Remove application list and update tabs to use buttons.
 
 1.2.0 - (February 15, 2018)
 ------------------